feat(role): add ProhibitedActionOnBaseRoleException to prevent updates and deletions of base roles

+++ src/routes/role/role.service.ts
@@ -1,7 +1,8 @@
 import { BadRequestException, Injectable } from '@nestjs/common'
-import { RoleAlreadyExistsException } from 'src/routes/role/role.error'
+import { ProhibitedActionOnBaseRoleException, RoleAlreadyExistsException } from 'src/routes/role/role.error'
 import { CreateRoleBodyType, GetRolesQueryType, UpdateRoleBodyType } from 'src/routes/role/role.model'
 import { RoleRepo } from 'src/routes/role/role.repo'
+import { RoleName } from 'src/shared/constants/role.constant'
 import { NotFoundRecordException } from 'src/shared/error'
 import { isNotFoundPrismaError, isUniqueConstraintPrismaError } from 'src/shared/helpers'
 
@@ -39,12 +40,20 @@ export class RoleService {
 
   async update({ id, data, updatedById }: { id: number; data: UpdateRoleBodyType; updatedById: number }) {
     try {
-      const role = await this.roleRepo.update({
+      const role = await this.roleRepo.findById(id)
+      if (!role) {
+        throw NotFoundRecordException
+      }
+      // Không cho bất kỳ ai cập nhật role ADMIN
+      if (role.name === RoleName.Admin) {
+        throw ProhibitedActionOnBaseRoleException
+      }
+      const updatedRole = await this.roleRepo.update({
         id,
         updatedById,
         data,
       })
-      return role
+      return updatedRole
     } catch (error) {
       if (isNotFoundPrismaError(error)) {
         throw NotFoundRecordException
@@ -61,6 +70,15 @@ export class RoleService {
 
   async delete({ id, deletedById }: { id: number; deletedById: number }) {
     try {
+      const role = await this.roleRepo.findById(id)
+      if (!role) {
+        throw NotFoundRecordException
+      }
+      // Không cho phép bất kỳ ai có thể xóa 3 role cơ bản này
+      const baseRoles: string[] = [RoleName.Admin, RoleName.Client, RoleName.Seller]
+      if (baseRoles.includes(role.name)) {
+        throw ProhibitedActionOnBaseRoleException
+      }
       await this.roleRepo.delete({
         id,
         deletedById,
